pyproject.toml
[project]
name = "market-insight"
version = "0.1.0"
description = ""
authors = [
    {name = "김준성",email = "gomshiki@gmail.com"}
]
requires-python = ">=3.14"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pyarrow (>=23.0.1,<24.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{include = "market_insight", from = "src"}]
